serializing throw works

+++ hardware/dart/inc/player.h
@@ -18,22 +18,25 @@
 /**
  * @brief OK - good throw just subtract, END - finish game player has 0 points, ERROR - player throw to much
  */
-enum ThrowStatus{
+enum ThrowStatus {
     ThrowStatus_OK, ThrowStatus_END, ThrowStatus_ERROR  
 };
 
 extern Dartboard dartboard;
 
+
+
+
 class Player{
     public:
-        uint32_t id;
+        uint16_t id;
         String nick;
         uint16_t points;
         uint8_t attemps = 0;
         Throw lastThrow = Throw(0,0);
 
         Player();
-        Player(const uint32_t &id, const String &nick, const uint16_t &points = 301, const uint8_t &attemps = 0);
+        Player(const uint16_t &id, const String &nick, const uint16_t &points = 301, const uint8_t &attemps = 0);
         Player &operator=(const Player &other);
